mysql.connector

+++ crane_monitoring_app/setup.sh
@@ -41,17 +41,31 @@ fi
 
 sudo apt update
 
-echo "\n\ninstalling MySQL Server..." 
-cd /tmp/
-mySqlPackege="mysql-apt-config_0.8.22-1_all.deb"
-wget "https://dev.mysql.com/get/$mySqlPackege"
-sudo dpkg -i $mySqlPackege
-#rm -f $mySqlPackege
-sudo apt update
-sudo apt-get install mysql-server
-#sudo nano /etc/mysql/mysql.conf.d/mysqld.cnf
-#bind-address=127.0.0.1
-systemctl restart mysql
+read -p "Install Mysql? 0 - no, 1 - yes: " installMysql
+if [ $installMysql == 0 ]; then
+  echo "Cntinue without installing MySQL"
+else
+  echo "\n\nInstalling MySQL Server..." 
+  cd /tmp/
+  mySqlPackege="mysql-apt-config_0.8.22-1_all.deb"
+  wget "https://dev.mysql.com/get/$mySqlPackege"
+  sudo dpkg -i $mySqlPackege
+  #rm -f $mySqlPackege
+  sudo apt update
+  sudo apt-get install mysql-server
+  #sudo nano /etc/mysql/mysql.conf.d/mysqld.cnf
+  #bind-address=127.0.0.1
+  systemctl restart mysql
+fi
+
+read -p "Install Mysql? 0 - no, 1 - yes: " installMysqlConnector
+if [ $installMysqlConnector == 0 ]; then
+  echo "Cntinue without installing mysql-connector-python"
+else
+  echo "\n\ninstalling mysql-connector-python" 
+  python3 pip install mysql-connector-python
+fi
+
 # sudo apt install gnome-control-center
 
 mkdir ~/developer
